Remove subclass loading and use base controller
