+++ modules/costs/app/views/admin/settings/time_entry_activities/reassign.html.erb
@@ -30,19 +30,19 @@ See COPYRIGHT and LICENSE files for more details.
 
 <%=
   render(Primer::OpenProject::PageHeader.new) do |header|
-    header.with_title { I18n.t(:label_new_time_entry_activity) }
-    header.with_description { I18n.t(:label_time_entry_activity_form_description) }
+    header.with_title { @enumeration.name }
+    header.with_description { t(:text_enumeration_destroy_question, count: @enumeration.objects_count) }
     header.with_breadcrumbs(
       [
         { href: admin_index_path, text: t("label_administration") },
         { href: admin_costs_settings_path, text: t(:project_module_costs) },
-        { href: admin_time_entry_activities_path, text: TimeEntryActivity.model_name.human(count: :other) },
-        I18n.t(:label_new_time_entry_activity)
+        { href: admin_settings_time_entry_activities_path, text: TimeEntryActivity.model_name.human(count: :other) },
+        @enumeration.name
       ]
     )
   end
 %>
 
-<%= primer_form_with(model: @enumeration, scope: :enumeration, url: url_for(action: :create)) do |f| %>
-  <%= render(Admin::Enumerations::ItemForm.new(f)) %>
+<%= primer_form_with(model: @enumeration, scope: :enumeration, url: url_for(action: :destroy, id: @enumeration), method: :delete) do |f| %>
+  <%= render(Admin::Enumerations::ReassignForm.new(f, other_enumerations: @other_enumerations)) %>
 <% end %>
